// packages/sdk/contracts-provider/service/src/factory/ContractsFactory.ts
export enum OpenZeppelinVersion {
  V4_9_6 = '4.9.6',
  V5_6_1 = '5.6.1',
}

export class ContractsFactory {
  /**
   * Retrieves an ERC20 wrapper implementation for the specified OpenZeppelin version
   */
  public static getERC20<TClient extends IBlockchainClient, TAddress extends IAddress>(params: {
    blockchainClient: TClient
    chainInfo: IChainInfo
    address: TAddress
    version?: OpenZeppelinVersion
  }) {
    const version = params.version ?? OpenZeppelinVersion.V5_6_1
    switch (version) {
      case OpenZeppelinVersion.V4_9_6:
        return new ERC20V4_9_6<TClient, TAddress>({
          blockchainClient: params.blockchainClient,
          chainInfo: params.chainInfo,
          address: params.address,
        })
      case OpenZeppelinVersion.V5_6_1:
      default:
        return new ERC20V5_6_1<TClient, TAddress>({
          blockchainClient: params.blockchainClient,
          chainInfo: params.chainInfo,
          address: params.address,
        })
    }
  }

  /**
   * Retrieves an ERC4626 wrapper implementation for the specified OpenZeppelin version
   */
  public static getERC4626<TClient extends IBlockchainClient, TAddress extends IAddress>(params: {
    blockchainClient: TClient
    chainInfo: IChainInfo
    address: TAddress
    version?: OpenZeppelinVersion
  }) {
    const version = params.version ?? OpenZeppelinVersion.V5_6_1
    switch (version) {
      case OpenZeppelinVersion.V4_9_6:
        return new ERC4626V4_9_6<TClient, TAddress>({
          blockchainClient: params.blockchainClient,
          chainInfo: params.chainInfo,
          address: params.address,
        })
      case OpenZeppelinVersion.V5_6_1:
      default:
        return new ERC4626V5_6_1<TClient, TAddress>({
          blockchainClient: params.blockchainClient,
          chainInfo: params.chainInfo,
          address: params.address,
        })
    }
  }

  /**
   * Retrieves an ERC1155 wrapper implementation for the specified OpenZeppelin version
   */
  public static getERC1155<TClient extends IBlockchainClient, TAddress extends IAddress>(params: {
    blockchainClient: TClient
    chainInfo: IChainInfo
    address: TAddress
    version?: OpenZeppelinVersion
  }) {
    const version = params.version ?? OpenZeppelinVersion.V5_6_1
    switch (version) {
      case OpenZeppelinVersion.V4_9_6:
        return new ERC1155V4_9_6<TClient, TAddress>({
          blockchainClient: params.blockchainClient,
          chainInfo: params.chainInfo,
          address: params.address,
        })
      case OpenZeppelinVersion.V5_6_1:
      default:
        return new ERC1155V5_6_1<TClient, TAddress>({
          blockchainClient: params.blockchainClient,
          chainInfo: params.chainInfo,
          address: params.address,
        })
    }
  }

  /**
   * Retrieves an ERC721 wrapper implementation for the specified OpenZeppelin version
   */
  public static getERC721<TClient extends IBlockchainClient, TAddress extends IAddress>(params: {
    blockchainClient: TClient
    chainInfo: IChainInfo
    address: TAddress
    version?: OpenZeppelinVersion
  }) {
    const version = params.version ?? OpenZeppelinVersion.V5_6_1
    switch (version) {
      case OpenZeppelinVersion.V4_9_6:
        return new ERC721V4_9_6<TClient, TAddress>({
          blockchainClient: params.blockchainClient,
          chainInfo: params.chainInfo,
          address: params.address,
        })
      case OpenZeppelinVersion.V5_6_1:
      default:
        return new ERC721V5_6_1<TClient, TAddress>({
          blockchainClient: params.blockchainClient,
          chainInfo: params.chainInfo,
          address: params.address,
        })
    }
  }
}